after Merge with current dev. Functions from SettingsNavigationHelper moved to ExternalActionHelper
