+++ Corona-Warn-App/src/test/java/de/rki/coronawarnapp/util/ExternalActionHelperTest.kt
@@ -1,26 +1,86 @@
 package de.rki.coronawarnapp.util
 
+import android.content.Context
+import android.content.pm.ApplicationInfo
 import androidx.fragment.app.Fragment
+import de.rki.coronawarnapp.CoronaWarnApplication
+import io.mockk.MockKAnnotations
 import io.mockk.Runs
 import io.mockk.every
+import io.mockk.impl.annotations.MockK
 import io.mockk.just
 import io.mockk.mockk
+import io.mockk.mockkObject
+import io.mockk.unmockkAll
 import io.mockk.verify
+import org.junit.After
+import org.junit.Before
 import org.junit.Test
 
-/**
- * ExternalActionHelper test.
- */
 class ExternalActionHelperTest {
 
-    /**
-     * Test activity called.
-     */
+    @MockK
+    private lateinit var context: Context
+
+    @MockK
+    private lateinit var applicationInfo: ApplicationInfo
+
+    @Before
+    fun setUp() {
+        MockKAnnotations.init(this)
+        mockkObject(CoronaWarnApplication)
+    }
+
     @Test
-    fun testCall() {
+    fun toConnectionsTest() {
+        every { context.startActivity(any()) } just Runs
+        ExternalActionHelper.toConnections(context = context)
+        verify(exactly = 1) { context.startActivity(any()) }
+    }
+
+    @Test
+    fun toNotificationsTest() {
+        every { context.startActivity(any()) } just Runs
+        every { context.packageName } returns "package_name"
+        every { context.applicationInfo } returns applicationInfo
+
+        ExternalActionHelper.toNotifications(context = context)
+        verify(exactly = 1) { context.startActivity(any()) }
+    }
+
+    @Test
+    fun toMainSettingsTest() {
+        every { context.startActivity(any()) } just Runs
+        ExternalActionHelper.toConnections(context = context)
+        verify(exactly = 1) { context.startActivity(any()) }
+    }
+
+    @Test
+    fun openUrlTest() {
         val fragment = mockk<Fragment>()
         every { fragment.startActivity(any()) } just Runs
-        ExternalActionHelper.call(fragment, "+77777777777")
+        ExternalActionHelper.openUrl(fragment = fragment, url = "url_path")
         verify(exactly = 1) { fragment.startActivity(any()) }
+    }
+
+    @Test
+    fun callTest() {
+        val fragment = mockk<Fragment>()
+        every { fragment.startActivity(any()) } just Runs
+        ExternalActionHelper.call(fragment = fragment, uri = "call_path")
+        verify(exactly = 1) { fragment.startActivity(any()) }
+    }
+
+    @Test
+    fun shareTextTest() {
+        val fragment = mockk<Fragment>()
+        every { fragment.startActivity(any()) } just Runs
+        ExternalActionHelper.shareText(fragment = fragment, text = "text", title = "title")
+        verify(exactly = 1) { fragment.startActivity(any()) }
+    }
+
+    @After
+    fun cleanUp() {
+        unmockkAll()
     }
 }
